+++ chaincode/model/contract_test.go
@@ -78,31 +78,3 @@ func TestRegisterModel(t *testing.T) {
 	_, err = contract.RegisterModel(ctx, wrapper)
 	assert.NoError(t, err)
 }
-
-func TestCanDisableModel(t *testing.T) {
-	contract := &SmartContract{}
-
-	mspid := "org"
-
-	wrapper, err := communication.Wrap(context.Background(), &asset.CanDisableModelParam{ModelKey: "uuid"})
-	assert.NoError(t, err)
-
-	ctx := new(ledger.MockTransactionContext)
-
-	service := getMockedService(ctx)
-	service.On("CanDisableModel", "uuid", mspid).Return(true, nil).Once()
-
-	stub := new(testHelper.MockedStub)
-	ctx.On("GetStub").Return(stub).Once()
-
-	stub.On("GetCreator").Return(testHelper.FakeTxCreator(t, mspid), nil).Once()
-
-	wrapped, err := contract.CanDisableModel(ctx, wrapper)
-	assert.NoError(t, err)
-
-	resp := new(asset.CanDisableModelResponse)
-	err = wrapped.Unwrap(resp)
-	assert.NoError(t, err)
-
-	assert.True(t, resp.CanDisable)
-}