+++ server/distributed/adapters/model.go
@@ -73,23 +73,6 @@ func (a *ModelAdapter) GetComputeTaskOutputModels(ctx context.Context, param *as
 	return response, nil
 }
 
-func (a *ModelAdapter) CanDisableModel(ctx context.Context, param *asset.CanDisableModelParam) (*asset.CanDisableModelResponse, error) {
-	invocator, err := interceptors.ExtractInvocator(ctx)
-	if err != nil {
-		return nil, err
-	}
-	method := "orchestrator.model:CanDisableModel"
-
-	response := new(asset.CanDisableModelResponse)
-
-	err = invocator.Call(ctx, method, param, response)
-	if err != nil {
-		return nil, err
-	}
-
-	return response, nil
-}
-
 func (a *ModelAdapter) RegisterModels(ctx context.Context, param *asset.RegisterModelsParam) (*asset.RegisterModelsResponse, error) {
 	Invocator, err := interceptors.ExtractInvocator(ctx)
 	if err != nil {